agentcarbon/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv

pytesseract
Pillow
opencv-python
paddlepaddle
paddleocr

transformers
torch
sentencepiece

pandas
prophet

qdrant-client
numpy
scikit-learn
